Enhance `ApplicationCommand`
Recommend to use `ApplicationCommand` instead of `GrailsApplicationCommand`
